Correcting schematic issue, adding preliminary board.
The previous power guarding method -- using the LT1910 to current sense
and also turn off the power line at the same time -- wasn't going to
work as its current sensing was way too small. Instead I've replaced
that part of the process with an actual current sensor to let the
microcontroller know when we've had an overcurrent condition.